Fix for laba 2

+++ src/main/java/com/gallery/catalog/dto/TransactionRequest.java
@@ -4,6 +4,7 @@ public record TransactionRequest(
     String userFullName,
     String galleryName,
     String paintingTitle,
-    Long price,
+    Double price,
     Integer year
-) {}
+) {
+}
